Create ServerStatus in Core
